@@ -95,12 +95,7 @@ public:
 
   inline void resize(size_t size) {
     if (size > builder.capacity()) grow(size);
-    while (builder.size() < size) {
-      builder.add(T());
-    }
-    while (builder.size() > size) {
-      builder.removeLast();
-    }
+    builder.resize(size);
   }
 
   inline void operator=(decltype(nullptr)) {
@@ -114,8 +109,12 @@ public:
   }
 
   inline void truncate(size_t size) {
-    while (builder.size() > size) {
-      builder.removeLast();
+    builder.truncate(size);
+  }
+
+  inline void reserve(size_t size) {
+    if (size > builder.capacity()) {
+      setCapacity(size);
     }
   }
 
@@ -126,11 +125,11 @@ private:
     setCapacity(kj::max(minCapacity, capacity() == 0 ? 4 : capacity() * 2));
   }
   void setCapacity(size_t newSize) {
-    ArrayBuilder<T> newBuilder = heapArrayBuilder<T>(newSize);
-    size_t moveCount = kj::min(newSize, builder.size());
-    for (size_t i = 0; i < moveCount; i++) {
-      newBuilder.add(kj::mv(builder[i]));
+    if (builder.size() > newSize) {
+      builder.truncate(newSize);
     }
+    ArrayBuilder<T> newBuilder = heapArrayBuilder<T>(newSize);
+    newBuilder.addAll(kj::mv(builder));
     builder = kj::mv(newBuilder);
   }
 };
